+++ editor/src/components/canvas/canvas-strategies/interaction-state.ts
@@ -22,7 +22,7 @@ import {
   CustomStrategyState,
   defaultCustomStrategyState,
 } from './canvas-strategy-types'
-import type { ReparentTarget } from './strategies/reparent-strategy-helpers'
+import type { ReparentTarget } from './strategies/reparent-helpers/reparent-strategy-helpers'
 
 export interface DragInteractionData {
   type: 'DRAG'
