+++ editor/src/components/canvas/canvas-strategies/strategies/absolute-reparent-strategy.tsx
@@ -21,8 +21,9 @@ import {
 import { InteractionSession, UpdatedPathMap } from '../interaction-state'
 import { absoluteMoveStrategy } from './absolute-move-strategy'
 import { honoursPropsPosition } from './absolute-utils'
-import { ifAllowedToReparent, isAllowedToReparent } from './reparent-helpers'
-import { getAbsoluteReparentPropertyChanges, ReparentTarget } from './reparent-strategy-helpers'
+import { ifAllowedToReparent, isAllowedToReparent } from './reparent-helpers/reparent-helpers'
+import { getAbsoluteReparentPropertyChanges } from './reparent-helpers/reparent-property-changes'
+import { ReparentTarget } from './reparent-helpers/reparent-strategy-helpers'
 import { getReparentOutcome, pathToReparent } from './reparent-utils'
 import { getDragTargets } from './shared-move-strategies-helpers'
