+++ editor/src/components/canvas/canvas-strategies/strategies/draw-to-insert-metastrategy.tsx
@@ -56,7 +56,7 @@ import {
 } from '../canvas-strategy-types'
 import { boundingArea, InteractionSession } from '../interaction-state'
 import { getApplicableReparentFactories } from './reparent-metastrategy'
-import { ReparentStrategy } from './reparent-strategy-helpers'
+import { ReparentStrategy } from './reparent-helpers/reparent-strategy-helpers'
 
 export const drawToInsertMetaStrategy: MetaCanvasStrategy = (
   canvasState: InteractionCanvasState,
