+++ editor/src/components/canvas/canvas-strategies/strategies/reparent-as-static-strategy.tsx
@@ -1,4 +1,17 @@
+import { MetadataUtils } from '../../../../core/model/element-metadata-utils'
+import * as EP from '../../../../core/shared/element-path'
+import { zeroCanvasRect } from '../../../../core/shared/math-utils'
 import { assertNever } from '../../../../core/shared/utils'
+import { absolute } from '../../../../utils/utils'
+import { CSSCursor } from '../../canvas-types'
+import { CanvasCommand } from '../../commands/commands'
+import { reorderElement } from '../../commands/reorder-element-command'
+import { setCursorCommand } from '../../commands/set-cursor-command'
+import { setElementsToRerenderCommand } from '../../commands/set-elements-to-rerender-command'
+import { showReorderIndicator } from '../../commands/show-reorder-indicator-command'
+import { updateHighlightedViews } from '../../commands/update-highlighted-views-command'
+import { updateSelectedViews } from '../../commands/update-selected-views-command'
+import { wildcardPatch } from '../../commands/wildcard-patch-command'
 import { ParentBounds } from '../../controls/parent-bounds'
 import { ParentOutlines } from '../../controls/parent-outlines'
 import {
@@ -11,11 +24,16 @@ import { CanvasStrategyFactory } from '../canvas-strategies'
 import {
   CanvasStrategy,
   controlWithProps,
+  emptyStrategyApplicationResult,
   getTargetPathsFromInteractionTarget,
   InteractionCanvasState,
+  StrategyApplicationResult,
 } from '../canvas-strategy-types'
 import { InteractionSession } from '../interaction-state'
-import { applyStaticReparent, ReparentTarget } from './reparent-strategy-helpers'
+import { ifAllowedToReparent } from './reparent-helpers/reparent-helpers'
+import { getStaticReparentPropertyChanges } from './reparent-helpers/reparent-property-changes'
+import { ReparentTarget } from './reparent-helpers/reparent-strategy-helpers'
+import { getReparentOutcome, pathToReparent } from './reparent-utils'
 import { getDragTargets } from './shared-move-strategies-helpers'
 
 export function baseReparentAsStaticStrategy(
@@ -97,4 +115,141 @@ function getIdAndNameOfReparentToStaticStrategy(targetLayout: 'flex' | 'flow'): 
     default:
       assertNever(targetLayout)
   }
+}
+
+export function applyStaticReparent(
+  canvasState: InteractionCanvasState,
+  interactionSession: InteractionSession,
+  reparentResult: ReparentTarget,
+): StrategyApplicationResult {
+  const selectedElements = getTargetPathsFromInteractionTarget(canvasState.interactionTarget)
+  const filteredSelectedElements = getDragTargets(selectedElements)
+
+  return ifAllowedToReparent(
+    canvasState,
+    canvasState.startingMetadata,
+    filteredSelectedElements,
+    () => {
+      if (
+        interactionSession.interactionData.type == 'DRAG' &&
+        interactionSession.interactionData.drag != null
+      ) {
+        if (reparentResult.shouldReparent && filteredSelectedElements.length === 1) {
+          const target = filteredSelectedElements[0]
+
+          const newIndex = reparentResult.newIndex
+          const newParent = reparentResult.newParent
+          const parentRect =
+            MetadataUtils.getFrameInCanvasCoords(newParent, canvasState.startingMetadata) ??
+            zeroCanvasRect
+
+          const siblingsOfTarget = MetadataUtils.getChildrenPaths(
+            canvasState.startingMetadata,
+            newParent,
+          )
+
+          const newParentADescendantOfCurrentParent = EP.isDescendantOfOrEqualTo(
+            newParent,
+            EP.parentPath(target),
+          )
+          // Reparent the element.
+          const outcomeResult = getReparentOutcome(
+            canvasState.builtInDependencies,
+            canvasState.projectContents,
+            canvasState.nodeModules,
+            canvasState.openFile,
+            pathToReparent(target),
+            newParent,
+            'always',
+          )
+
+          if (outcomeResult != null) {
+            const { commands: reparentCommands, newPath } = outcomeResult
+
+            const targetMetadata = MetadataUtils.findElementByElementPath(
+              canvasState.startingMetadata,
+              target,
+            )
+
+            // Strip the `position`, positional and dimension properties.
+            const propertyChangeCommands = getStaticReparentPropertyChanges(
+              newPath,
+              targetMetadata?.specialSizeMeasurements.position ?? null,
+              targetMetadata?.specialSizeMeasurements.display ?? null,
+              reparentResult.shouldConvertToInline,
+            )
+
+            const commandsBeforeReorder = [
+              ...reparentCommands,
+              updateSelectedViews('always', [newPath]),
+            ]
+
+            const commandsAfterReorder = [
+              ...propertyChangeCommands,
+              setElementsToRerenderCommand([target, newPath]),
+              updateHighlightedViews('mid-interaction', []),
+              setCursorCommand(CSSCursor.Move),
+            ]
+
+            function midInteractionCommandsForTarget(shouldReorder: boolean): Array<CanvasCommand> {
+              const commonPatches = [
+                wildcardPatch('mid-interaction', {
+                  canvas: { controls: { parentHighlightPaths: { $set: [newParent] } } },
+                }),
+                newParentADescendantOfCurrentParent
+                  ? wildcardPatch('mid-interaction', {
+                      hiddenInstances: { $push: [target] },
+                    })
+                  : wildcardPatch('mid-interaction', {
+                      displayNoneInstances: { $push: [target] },
+                    }),
+              ]
+              if (shouldReorder) {
+                return [
+                  ...commonPatches,
+                  showReorderIndicator(newParent, newIndex),
+                  wildcardPatch('mid-interaction', {
+                    displayNoneInstances: { $push: [newPath] },
+                  }),
+                ]
+              } else {
+                return commonPatches
+              }
+            }
+
+            let interactionFinishCommands: Array<CanvasCommand>
+            let midInteractionCommands: Array<CanvasCommand>
+
+            if (reparentResult.shouldReorder && siblingsOfTarget.length > 0) {
+              midInteractionCommands = midInteractionCommandsForTarget(reparentResult.shouldReorder)
+
+              interactionFinishCommands = [
+                ...commandsBeforeReorder,
+                reorderElement('always', newPath, absolute(newIndex)),
+                ...commandsAfterReorder,
+              ]
+            } else {
+              if (parentRect != null) {
+                midInteractionCommands = midInteractionCommandsForTarget(
+                  reparentResult.shouldReorder,
+                )
+              } else {
+                // this should be an error because parentRect should never be null
+                midInteractionCommands = []
+              }
+
+              interactionFinishCommands = [...commandsBeforeReorder, ...commandsAfterReorder]
+            }
+
+            return {
+              commands: [...midInteractionCommands, ...interactionFinishCommands],
+              customStatePatch: {},
+              status: 'success',
+            }
+          }
+        }
+      }
+      return emptyStrategyApplicationResult
+    },
+  )
 }
