+++ editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-property-changes.ts
@@ -1,0 +1,203 @@
+import { isHorizontalPoint } from 'utopia-api/core'
+import { getLayoutProperty } from '../../../../../core/layout/getLayoutProperty'
+import {
+  framePointForPinnedProp,
+  LayoutPinnedProp,
+} from '../../../../../core/layout/layout-helpers-new'
+import { MetadataUtils } from '../../../../../core/model/element-metadata-utils'
+import { mapDropNulls } from '../../../../../core/shared/array-utils'
+import { isRight, right } from '../../../../../core/shared/either'
+import * as EP from '../../../../../core/shared/element-path'
+import { ElementInstanceMetadataMap, JSXElement } from '../../../../../core/shared/element-template'
+import {
+  canvasPoint,
+  pointDifference,
+  roundPointToNearestHalf,
+} from '../../../../../core/shared/math-utils'
+import { ElementPath, PropertyPath } from '../../../../../core/shared/project-file-types'
+import * as PP from '../../../../../core/shared/property-path'
+import { ProjectContentTreeRoot } from '../../../../assets'
+import { getElementFromProjectContents } from '../../../../editor/store/editor-state'
+import { CSSPosition } from '../../../../inspector/common/css-utils'
+import { stylePropPathMappingFn } from '../../../../inspector/common/property-path-hooks'
+import {
+  AdjustCssLengthProperty,
+  adjustCssLengthProperty,
+} from '../../../commands/adjust-css-length-command'
+import { CanvasCommand } from '../../../commands/commands'
+import { deleteProperties } from '../../../commands/delete-properties-command'
+import { setProperty } from '../../../commands/set-property-command'
+import {
+  getOptionalCommandToConvertDisplayInlineBlock,
+  singleAxisAutoLayoutContainerDirections,
+} from '../flow-reorder-helpers'
+import { ReparentStrategy } from './reparent-strategy-helpers'
+
+const propertiesToRemove: Array<PropertyPath> = [
+  PP.create(['style', 'left']),
+  PP.create(['style', 'top']),
+  PP.create(['style', 'right']),
+  PP.create(['style', 'bottom']),
+]
+
+export function getAbsoluteReparentPropertyChanges(
+  target: ElementPath,
+  newParent: ElementPath,
+  targetStartingMetadata: ElementInstanceMetadataMap,
+  newParentStartingMetadata: ElementInstanceMetadataMap,
+  projectContents: ProjectContentTreeRoot,
+  openFile: string | null | undefined,
+): Array<AdjustCssLengthProperty> {
+  const element: JSXElement | null = getElementFromProjectContents(
+    target,
+    projectContents,
+    openFile,
+  )
+
+  if (element == null) {
+    return []
+  }
+
+  const currentParentContentBox = MetadataUtils.getParentCoordinateSystemBounds(
+    EP.parentPath(target),
+    targetStartingMetadata,
+  )
+  const newParentContentBox = MetadataUtils.getParentCoordinateSystemBounds(
+    newParent,
+    newParentStartingMetadata,
+  )
+
+  const offsetTL = roundPointToNearestHalf(
+    pointDifference(newParentContentBox, currentParentContentBox),
+  )
+  const offsetBR = roundPointToNearestHalf(
+    pointDifference(
+      canvasPoint({
+        x: currentParentContentBox.x + currentParentContentBox.width,
+        y: currentParentContentBox.y + currentParentContentBox.height,
+      }),
+      canvasPoint({
+        x: newParentContentBox.x + newParentContentBox.width,
+        y: newParentContentBox.y + newParentContentBox.height,
+      }),
+    ),
+  )
+
+  const createAdjustCssLengthProperty = (
+    pin: LayoutPinnedProp,
+    newValue: number,
+    parentDimension: number | undefined,
+  ): AdjustCssLengthProperty | null => {
+    const value = getLayoutProperty(pin, right(element.props), ['style'])
+    if (isRight(value) && value.value != null) {
+      // TODO what to do about missing properties?
+      return adjustCssLengthProperty(
+        'always',
+        target,
+        stylePropPathMappingFn(pin, ['style']),
+        newValue,
+        parentDimension,
+        true,
+      )
+    } else {
+      return null
+    }
+  }
+
+  const newParentFrame = MetadataUtils.getFrameInCanvasCoords(newParent, newParentStartingMetadata)
+
+  return [
+    ...mapDropNulls(
+      (pin) => {
+        const horizontal = isHorizontalPoint(framePointForPinnedProp(pin))
+        return createAdjustCssLengthProperty(
+          pin,
+          horizontal ? offsetTL.x : offsetTL.y,
+          horizontal ? newParentFrame?.width : newParentFrame?.height,
+        )
+      },
+      ['top', 'left'] as const,
+    ),
+    ...mapDropNulls(
+      (pin) => {
+        const horizontal = isHorizontalPoint(framePointForPinnedProp(pin))
+        return createAdjustCssLengthProperty(
+          pin,
+          horizontal ? offsetBR.x : offsetBR.y,
+          horizontal ? newParentFrame?.width : newParentFrame?.height,
+        )
+      },
+      ['bottom', 'right'] as const,
+    ),
+  ]
+}
+
+export function getStaticReparentPropertyChanges(
+  newPath: ElementPath,
+  targetOriginalStylePosition: CSSPosition | null,
+  targetOriginalDisplayProp: string | null,
+  convertToInline: 'row' | 'column' | 'do-not-convert',
+): Array<CanvasCommand> {
+  const optionalInlineConversionCommand = getOptionalCommandToConvertDisplayInlineBlock(
+    newPath,
+    targetOriginalDisplayProp,
+    convertToInline,
+  )
+
+  if (targetOriginalStylePosition !== 'absolute' && targetOriginalStylePosition !== 'relative') {
+    return [
+      ...optionalInlineConversionCommand,
+      deleteProperties('always', newPath, propertiesToRemove),
+    ]
+  }
+
+  return [
+    ...optionalInlineConversionCommand,
+    deleteProperties('always', newPath, [...propertiesToRemove, PP.create(['style', 'position'])]),
+    setProperty('always', newPath, PP.create(['style', 'contain']), 'layout'),
+  ]
+}
+
+export function getReparentPropertyChanges(
+  reparentStrategy: ReparentStrategy,
+  target: ElementPath,
+  newParent: ElementPath,
+  targetStartingMetadata: ElementInstanceMetadataMap,
+  newParentStartingMetadata: ElementInstanceMetadataMap,
+  projectContents: ProjectContentTreeRoot,
+  openFile: string | null | undefined,
+  targetOriginalStylePosition: CSSPosition | null,
+  targetOriginalDisplayProp: string | null,
+): Array<CanvasCommand> {
+  switch (reparentStrategy) {
+    case 'REPARENT_AS_ABSOLUTE':
+      return getAbsoluteReparentPropertyChanges(
+        target,
+        newParent,
+        targetStartingMetadata,
+        newParentStartingMetadata,
+        projectContents,
+        openFile,
+      )
+    case 'REPARENT_AS_STATIC':
+      const newPath = EP.appendToPath(newParent, EP.toUid(target))
+      const directions = singleAxisAutoLayoutContainerDirections(
+        newParent,
+        newParentStartingMetadata,
+      )
+
+      const convertDisplayInline =
+        directions === 'non-single-axis-autolayout' ||
+        directions.direction == null ||
+        directions.flexOrFlow === 'flex'
+          ? 'do-not-convert'
+          : directions.direction
+
+      return getStaticReparentPropertyChanges(
+        newPath,
+        targetOriginalStylePosition,
+        targetOriginalDisplayProp,
+        convertDisplayInline,
+      )
+  }
+}
