+++ editor/src/components/canvas/commands/show-reorder-indicator-command.ts
@@ -11,7 +11,7 @@ import { singleAxisAutoLayoutContainerDirections } from '../canvas-strategies/st
 import {
   getSiblingMidPointPosition,
   siblingAndPseudoPositions,
-} from '../canvas-strategies/strategies/reparent-strategy-helpers'
+} from '../canvas-strategies/strategies/reparent-helpers/reparent-strategy-sibling-position-helpers'
 import type { BaseCommand, CommandFunction, WhenToRun } from './commands'
 
 export interface ShowReorderIndicator extends BaseCommand {
